Add two function

+++ src/main/java/com/katalogus/project/model/StudentStatistic.java
@@ -17,6 +17,6 @@ public class StudentStatistic {
 
     private String neptunCode;
 
-    private HashMap<String, Double> percentages = new HashMap<>();
+    private HashMap<String, Integer> percentages = new HashMap<>();
 
 }
